requirements.txt
dependency_injector==4.32.3
PyYAML==5.4.1
pyserial==3.5
paho-mqtt==1.5.1
asyncio==3.4.3
packaging==20.9
dacite==1.6.0
aiofiles==0.7.0